Support for scalable beanfactory
